Move LocalDateTimeAmbiguityStatus enum to LocalDateTimeInfo
