+++ src/main/java/teammates/ui/webapi/output/LocalDateTimeInfo.java
@@ -1,13 +1,16 @@
 package teammates.ui.webapi.output;
 
-import teammates.common.util.TimeHelper;
+import java.time.LocalDateTime;
+import java.time.ZoneId;
+import java.time.ZoneOffset;
+import java.util.List;
 
 /**
  * The API output format of a {@code LocalDateTimeInfo} to hold information for resolving DST overlaps/gaps.
  */
 public class LocalDateTimeInfo extends ApiOutput {
     private long resolvedTimestamp;
-    private TimeHelper.LocalDateTimeAmbiguityStatus resolvedStatus;
+    private LocalDateTimeAmbiguityStatus resolvedStatus;
 
     private Long earlierInterpretationTimestamp;
     private Long laterInterpretationTimestamp;
@@ -18,7 +21,7 @@ public class LocalDateTimeInfo extends ApiOutput {
     public static LocalDateTimeInfo unambiguous(long resolvedTimestamp) {
         LocalDateTimeInfo localDateTimeInfo = new LocalDateTimeInfo();
 
-        localDateTimeInfo.resolvedStatus = TimeHelper.LocalDateTimeAmbiguityStatus.UNAMBIGUOUS;
+        localDateTimeInfo.resolvedStatus = LocalDateTimeAmbiguityStatus.UNAMBIGUOUS;
         localDateTimeInfo.resolvedTimestamp = resolvedTimestamp;
 
         return localDateTimeInfo;
@@ -30,7 +33,7 @@ public class LocalDateTimeInfo extends ApiOutput {
     public static LocalDateTimeInfo gap(long resolvedTimestamp) {
         LocalDateTimeInfo localDateTimeInfo = new LocalDateTimeInfo();
 
-        localDateTimeInfo.resolvedStatus = TimeHelper.LocalDateTimeAmbiguityStatus.GAP;
+        localDateTimeInfo.resolvedStatus = LocalDateTimeAmbiguityStatus.GAP;
         localDateTimeInfo.resolvedTimestamp = resolvedTimestamp;
 
         return localDateTimeInfo;
@@ -43,7 +46,7 @@ public class LocalDateTimeInfo extends ApiOutput {
                                             long earlierInterpretationTimestamp, long laterInterpretationTimestamp) {
         LocalDateTimeInfo localDateTimeInfo = new LocalDateTimeInfo();
 
-        localDateTimeInfo.resolvedStatus = TimeHelper.LocalDateTimeAmbiguityStatus.OVERLAP;
+        localDateTimeInfo.resolvedStatus = LocalDateTimeAmbiguityStatus.OVERLAP;
         localDateTimeInfo.resolvedTimestamp = resolvedTimestamp;
 
         localDateTimeInfo.earlierInterpretationTimestamp = earlierInterpretationTimestamp;
@@ -56,7 +59,7 @@ public class LocalDateTimeInfo extends ApiOutput {
         return resolvedTimestamp;
     }
 
-    public TimeHelper.LocalDateTimeAmbiguityStatus getResolvedStatus() {
+    public LocalDateTimeAmbiguityStatus getResolvedStatus() {
         return resolvedStatus;
     }
 
@@ -66,5 +69,47 @@ public class LocalDateTimeInfo extends ApiOutput {
 
     public Long getLaterInterpretationTimestamp() {
         return laterInterpretationTimestamp;
+    }
+
+    /**
+     * Represents the ambiguity status for a {@link LocalDateTime} at a given time {@code zone},
+     * brought about by Daylight Saving Time (DST).
+     */
+    public enum LocalDateTimeAmbiguityStatus {
+        /**
+         * The local date time can be unambiguously resolved to a single instant.
+         * It has only one valid interpretation.
+         */
+        UNAMBIGUOUS,
+
+        /**
+         * The local date time falls within the gap period when clocks spring forward at the start of DST.
+         * Strictly speaking, it is non-existent, and needs to be readjusted to be valid.
+         */
+        GAP,
+
+        /**
+         * The local date time falls within the overlap period when clocks fall back at the end of DST.
+         * It has more than one valid interpretation.
+         */
+        OVERLAP;
+
+        /**
+         * Gets the ambiguity status for a {@link LocalDateTime} at a given time {@code zone}.
+         */
+        public static LocalDateTimeAmbiguityStatus of(LocalDateTime localDateTime, ZoneId zone) {
+            if (localDateTime == null || zone == null) {
+                return null;
+            }
+
+            List<ZoneOffset> offsets = zone.getRules().getValidOffsets(localDateTime);
+            if (offsets.size() == 1) {
+                return UNAMBIGUOUS;
+            }
+            if (offsets.isEmpty()) {
+                return GAP;
+            }
+            return OVERLAP;
+        }
     }
 }
