+++ src/test/java/teammates/test/cases/webapi/GetLocalDateTimeInfoActionTest.java
@@ -4,7 +4,6 @@ import org.testng.annotations.Test;
 
 import teammates.common.datatransfer.attributes.InstructorAttributes;
 import teammates.common.util.Const;
-import teammates.common.util.TimeHelper;
 import teammates.ui.webapi.action.GetLocalDateTimeInfoAction;
 import teammates.ui.webapi.action.JsonResult;
 import teammates.ui.webapi.output.LocalDateTimeInfo;
@@ -54,7 +53,7 @@ public class GetLocalDateTimeInfoActionTest extends BaseActionTest<GetLocalDateT
         JsonResult r = getJsonResult(a);
         LocalDateTimeInfo localDateTimeInfo = (LocalDateTimeInfo) r.getOutput();
 
-        assertEquals(TimeHelper.LocalDateTimeAmbiguityStatus.UNAMBIGUOUS, localDateTimeInfo.getResolvedStatus());
+        assertEquals(LocalDateTimeInfo.LocalDateTimeAmbiguityStatus.UNAMBIGUOUS, localDateTimeInfo.getResolvedStatus());
         assertEquals(1543669200000L, localDateTimeInfo.getResolvedTimestamp());
         assertNull(localDateTimeInfo.getEarlierInterpretationTimestamp());
         assertNull(localDateTimeInfo.getLaterInterpretationTimestamp());
@@ -73,7 +72,7 @@ public class GetLocalDateTimeInfoActionTest extends BaseActionTest<GetLocalDateT
         r = getJsonResult(a);
         localDateTimeInfo = (LocalDateTimeInfo) r.getOutput();
 
-        assertEquals(TimeHelper.LocalDateTimeAmbiguityStatus.GAP, localDateTimeInfo.getResolvedStatus());
+        assertEquals(LocalDateTimeInfo.LocalDateTimeAmbiguityStatus.GAP, localDateTimeInfo.getResolvedStatus());
         assertEquals(1332639000000L, localDateTimeInfo.getResolvedTimestamp());
         assertNull(localDateTimeInfo.getEarlierInterpretationTimestamp());
         assertNull(localDateTimeInfo.getLaterInterpretationTimestamp());
@@ -92,7 +91,7 @@ public class GetLocalDateTimeInfoActionTest extends BaseActionTest<GetLocalDateT
         r = getJsonResult(a);
         localDateTimeInfo = (LocalDateTimeInfo) r.getOutput();
 
-        assertEquals(TimeHelper.LocalDateTimeAmbiguityStatus.OVERLAP, localDateTimeInfo.getResolvedStatus());
+        assertEquals(LocalDateTimeInfo.LocalDateTimeAmbiguityStatus.OVERLAP, localDateTimeInfo.getResolvedStatus());
         assertEquals(1351384200000L, localDateTimeInfo.getResolvedTimestamp());
         assertEquals(1351384200000L, localDateTimeInfo.getEarlierInterpretationTimestamp().longValue());
         assertEquals(1351387800000L, localDateTimeInfo.getLaterInterpretationTimestamp().longValue());
